services/messaging/requirements.txt
fastapi
uvicorn
better-profanity
supabase
python-dotenv
